refactor: extract admin tag formatters module
